scrape-traffic/requirements.txt
dpkt
pandas==1.4.2
psutil==5.8.0
python-dotenv==1.0.0
